client opencv updated to 4.4.0
- android client library of opencv has been updated to latest 4.4.0 version
-- includes the needed videowriter class of functions

+++ client_android/opencv/src/org/opencv/core/TermCriteria.java
@@ -1,57 +1,6 @@
 package org.opencv.core;
 
-/**
- * <p>class CV_EXPORTS TermCriteria <code></p>
- *
- * <p>// C++ code:</p>
- *
- *
- * <p>public:</p>
- *
- * <p>enum</p>
- *
- *
- * <p>COUNT=1, //!< the maximum number of iterations or elements to compute</p>
- *
- * <p>MAX_ITER=COUNT, //!< ditto</p>
- *
- * <p>EPS=2 //!< the desired accuracy or change in parameters at which the
- * iterative algorithm stops</p>
- *
- * <p>};</p>
- *
- * <p>//! default constructor</p>
- *
- * <p>TermCriteria();</p>
- *
- * <p>//! full constructor</p>
- *
- * <p>TermCriteria(int type, int maxCount, double epsilon);</p>
- *
- * <p>//! conversion from CvTermCriteria</p>
- *
- * <p>TermCriteria(const CvTermCriteria& criteria);</p>
- *
- * <p>//! conversion to CvTermCriteria</p>
- *
- * <p>operator CvTermCriteria() const;</p>
- *
- * <p>int type; //!< the type of termination criteria: COUNT, EPS or COUNT + EPS</p>
- *
- * <p>int maxCount; // the maximum number of iterations/elements</p>
- *
- * <p>double epsilon; // the desired accuracy</p>
- *
- * <p>};</p>
- *
- * <p>The class defining termination criteria for iterative algorithms. You can
- * initialize it by default constructor and then override any parameters, or the
- * structure may be fully initialized using the advanced variant of the
- * constructor.
- * </code></p>
- *
- * @see <a href="http://docs.opencv.org/modules/core/doc/basic_structures.html#termcriteria">org.opencv.core.TermCriteria</a>
- */
+//javadoc:TermCriteria
 public class TermCriteria {
 
     /**
@@ -102,7 +51,7 @@ public class TermCriteria {
         if (vals != null) {
             type = vals.length > 0 ? (int) vals[0] : 0;
             maxCount = vals.length > 1 ? (int) vals[1] : 0;
-            epsilon = vals.length > 2 ? vals[2] : 0;
+            epsilon = vals.length > 2 ? (double) vals[2] : 0;
         } else {
             type = 0;
             maxCount = 0;
